Move logging to blender.com.gltf2_io_debug
